@@ -7,23 +7,13 @@ class Config
     const CONFIG_KEY_MAGENTO_BOOTSTRAP = 'bootstrap';
     const CONFIG_KEY_SERVICES = 'services';
 
-    /**
-     * @var string
-     */
-    private $magentoBootstrapPath;
+    private string $magentoBootstrapPath;
+    private string $servicesPath;
 
-    /**
-     * @var string
-     */
-    private $servicesPath;
-
-    /**
-     * @param string[] $config
-     */
     public function __construct(array $config)
     {
-        $this->magentoBootstrapPath = $config[self::CONFIG_KEY_MAGENTO_BOOTSTRAP];
-        $this->servicesPath = $config[self::CONFIG_KEY_SERVICES];
+        $this->magentoBootstrapPath = (string) $config[self::CONFIG_KEY_MAGENTO_BOOTSTRAP];
+        $this->servicesPath = (string) $config[self::CONFIG_KEY_SERVICES];
     }
 
     public function getMagentoBootstrapPath(): string
